ref: Prepare store for SSR
